rename: domain 패키지로 menu 및 menuList 이동
