+++ src/main/java/com/ourMenu/backend/domain/menu/dao/MenuRepository.java
@@ -1,6 +1,6 @@
-package com.ourMenu.backend.menu.dao;
+package com.ourMenu.backend.domain.menu.dao;
 
-import com.ourMenu.backend.menu.domain.Menu;
+import com.ourMenu.backend.domain.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
