+++ src/main/java/com/ourMenu/backend/domain/menulist/dao/MenuListRepository.java
@@ -1,6 +1,6 @@
-package com.ourMenu.backend.menulist.dao;
+package com.ourMenu.backend.domain.menulist.dao;
 
-import com.ourMenu.backend.menulist.domain.MenuList;
+import com.ourMenu.backend.domain.menulist.domain.MenuList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuListRepository extends JpaRepository<MenuList, Long> {
