+++ src/main/java/com/ourMenu/backend/domain/menulist/domain/MenuListStatus.java
@@ -1,4 +1,4 @@
-package com.ourMenu.backend.menulist.domain;
+package com.ourMenu.backend.domain.menulist.domain;
 
 public enum MenuListStatus {
     CREATED,
